third_party/blink/web_tests/platform/mac-mac10.15/plugins/refcount-leaks-expected.txt

Test that we can get an NPObject returned through a method on an NPAPI Object.

Prints "SUCCESS" on success, "FAILURE" on failure. 

--- num test objects:
countAfterCreate == countOrig + 3? PASS
countOrig == countAfterGC? PASS
countOrig == countAfterGC2? PASS

SUCCESS
